refactor: make `FileProgress` inherit from `DequeProgress`
